requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.11.1
pandas>=1.5.0
Pillow>=9.3.0
requests>=2.28.1